Refactor: SubHeader 파일 이동, 주석 제거, 경로 변경
